backend/requirements.txt
Flask==3.1.0
Flask-CORS==5.0.1
PyJWT==2.10.1
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
marshmallow==3.26.1
Flask-JWT-Extended==4.7.1
requests==2.31.0
google-generativeai==0.8.5
FLask-Limiter==3.12
bcrypt==4.3.0
marshmallow==3.26.1
psycopg2-binary==2.9.10

